feat: refactor Koin database setup for iOS and Android, update color handling in UI components
