--- composeApp/src/androidMain/kotlin/com/debanshu/xcalendar/Platform.android.kt
@@ -1,9 +1,0 @@
-package com.debanshu.xcalendar
-
-import android.os.Build
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
