+++ composeApp/src/androidMain/kotlin/com/debanshu/xcalendar/di/Koin.android.kt
@@ -1,8 +1,19 @@
 package com.debanshu.xcalendar.di
 
-import org.koin.core.annotation.ComponentScan
-import org.koin.core.annotation.Module
+import android.content.Context
+import androidx.room.Room
+import androidx.sqlite.driver.bundled.BundledSQLiteDriver
+import com.debanshu.xcalendar.data.localDataSource.AppDatabase
+import kotlinx.coroutines.Dispatchers
+import org.koin.mp.KoinPlatform
 
-@Module
-@ComponentScan("com.debanshu.xcalendar.platform")
-actual class NativeModule
+actual fun getDatabase(): AppDatabase {
+    val context = KoinPlatform.getKoin().get<Context>()
+    return Room.databaseBuilder<AppDatabase>(
+        context,
+        context.getDatabasePath("calendar.db").absolutePath
+    )
+        .setDriver(BundledSQLiteDriver())
+        .setQueryCoroutineContext(Dispatchers.IO)
+        .build()
+}
