--- composeApp/src/commonMain/kotlin/com/debanshu/xcalendar/Greeting.kt
@@ -1,9 +1,0 @@
-package com.debanshu.xcalendar
-
-class Greeting {
-    private val platform = getPlatform()
-
-    fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
