--- composeApp/src/commonMain/kotlin/com/debanshu/xcalendar/Platform.kt
@@ -1,7 +1,0 @@
-package com.debanshu.xcalendar
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
