--- composeApp/src/commonMain/kotlin/com/debanshu/xcalendar/platform/DatabaseComponent.kt
@@ -1,7 +1,0 @@
-package com.debanshu.xcalendar.platform
-
-import com.debanshu.xcalendar.data.localDataSource.AppDatabase
-
-expect class DatabaseComponent {
-    fun getDatabase() : AppDatabase
-}
