+++ composeApp/src/iosMain/kotlin/com/debanshu/xcalendar/di/Koin.ios.kt
@@ -1,8 +1,32 @@
 package com.debanshu.xcalendar.di
 
-import org.koin.core.annotation.ComponentScan
-import org.koin.core.annotation.Module
+import androidx.room.Room
+import androidx.sqlite.driver.bundled.BundledSQLiteDriver
+import com.debanshu.xcalendar.data.localDataSource.AppDatabase
+import kotlinx.cinterop.ExperimentalForeignApi
+import kotlinx.coroutines.Dispatchers
+import platform.Foundation.NSDocumentDirectory
+import platform.Foundation.NSFileManager
+import platform.Foundation.NSUserDomainMask
 
-@Module
-@ComponentScan("com.debanshu.xcalendar.platform")
-actual class NativeModule
+actual fun getDatabase(): AppDatabase {
+    val dbFile = documentDirectory() + "/calendar.db"
+    return Room.databaseBuilder<AppDatabase>(
+        name = dbFile,
+    )
+        .setDriver(BundledSQLiteDriver())
+        .setQueryCoroutineContext(Dispatchers.Default)
+        .build()
+}
+
+@OptIn(ExperimentalForeignApi::class)
+private fun documentDirectory(): String {
+    val documentDirectory = NSFileManager.defaultManager.URLForDirectory(
+        directory = NSDocumentDirectory,
+        inDomain = NSUserDomainMask,
+        appropriateForURL = null,
+        create = false,
+        error = null,
+    )
+    return requireNotNull(documentDirectory?.path)
+}
