+++ gradle/libs.versions.toml
@@ -10,7 +10,7 @@ compose-multiplatform = "1.8.0-alpha03"
 kotlin = "2.1.10"
 room_version = "2.7.0-rc01"
 ksp = "2.1.10-1.0.31"
-sqlite = "2.5.0-SNAPSHOT"
+sqlite = "2.5.0-rc01"
 ktor-bom = "3.1.1"
 koin-bom = "4.1.0-Beta5"
 koin-annotations-bom = "2.0.0"
